components/pod-components/lymphocyte/requirements.txt
requests==2.24.0
redis==3.5.3
pymongo==3.11.1
python-dateutil